fix Contractor part + Model layer container Inheritance

+++ src/ModelLayer/Box.java
@@ -18,7 +18,7 @@ public abstract class Box {
         return height;
     }
 
-    public void setHeight(int height) {
+    public void setHeight(double height) {
         this.height = height;
     }
 
@@ -26,7 +26,7 @@ public abstract class Box {
         return length;
     }
 
-    public void setLength(int length) {
+    public void setLength(double length) {
         this.length = length;
     }
 
@@ -34,7 +34,7 @@ public abstract class Box {
         return width;
     }
 
-    public void setWidth(int width) {
+    public void setWidth(double width) {
         this.width = width;
     }
 }
